+++ lib/DevStandard/src/main/java/dev/standard/catalog/Config.java
@@ -100,6 +100,11 @@ final class Config {
         sDevUtilsCatelogMap.put(".DevOther", "第三方库封装、以及部分特殊工具类等, 方便 copy 封装类使用");
         sDevUtilsCatelogMap.put(".DevStandard", "项目规范统一检测、生成替换等");
         sDevUtilsCatelogMap.put(".DevWidget", "部分自定义 View 功能、效果");
+        sDevUtilsCatelogMap.put(".Environment", "环境配置切换库");
+        sDevUtilsCatelogMap.put(".Environment.DevEnvironment", "环境切换可视化 UI 操作");
+        sDevUtilsCatelogMap.put(".Environment.DevEnvironmentBase", "注解类、实体类、监听事件等通用基础");
+        sDevUtilsCatelogMap.put(".Environment.DevEnvironmentCompiler", "Debug ( 打包 / 编译 ) 生成实现代码");
+        sDevUtilsCatelogMap.put(".Environment.DevEnvironmentCompilerRelease", "Release ( 打包 / 编译 ) 生成实现代码");
 
         // =========================
         // = DevUtils Lib 忽略目录 =